setupscript generation: update compiler handling
The compiler is not a normal dependency of the bundle package, and so has to have its own logic to set it up in the setup script. Previously, this was done by adding the spec of the compiler to the list of dependencies - which only works for spack-installed compilers. This change adds the compiler paths directly to the environment changes (similar to what is done when setting up the build environment in spack) and should so also work with external compilers.