Adding package BDSim, default version v1.6.0 (#388)
* Adding package BDSim, default version v1.6.0

Beam Delivery Simulation (BDSIM) is a C++ program that utilises the Geant4 toolkit
to simulate both the transport of particles in an accelerator and their interaction
with the accelerator material. BDSIM is capable of simulating a wide variety of
accelerator components, magnets and processes, e.g. sychrotron radiation, with Geant4
geometry dynamically built based on a text input file. Thick lens accelerator tracking
routines are provided for fast accurate tracking in a vacuum.

* Use latest Geant4 for versions more recent than 1.6.0

* Update packages/bdsim/package.py

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>

* Update packages/bdsim/package.py

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>

* Update packages/bdsim/package.py

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>

* Remove bdsim recipe (moved upstream)

* Tidy package list in key4hep-stack

* Add bdsim to key4hep-stack

Co-authored-by: Valentin Volkl <valentin.volkl@cern.ch>